Fix attachement/gridfs sync

+++ mica-core/src/main/java/org/obiba/mica/file/service/FileSystemService.java
@@ -10,28 +10,11 @@
 
 package org.obiba.mica.file.service;
 
-import static java.util.stream.Collectors.toList;
-
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Paths;
-import java.time.LocalDateTime;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.locks.ReentrantLock;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipOutputStream;
-
-import javax.annotation.Nullable;
-import javax.inject.Inject;
-import javax.validation.constraints.NotNull;
-
+import com.google.common.base.Strings;
+import com.google.common.base.Throwables;
+import com.google.common.collect.Maps;
+import com.google.common.eventbus.EventBus;
+import com.google.common.eventbus.Subscribe;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.math3.util.Pair;
 import org.apache.shiro.SecurityUtils;
@@ -49,11 +32,7 @@ import org.obiba.mica.dataset.domain.HarmonizationDataset;
 import org.obiba.mica.dataset.event.DatasetPublishedEvent;
 import org.obiba.mica.dataset.event.DatasetUnpublishedEvent;
 import org.obiba.mica.dataset.event.DatasetUpdatedEvent;
-import org.obiba.mica.file.Attachment;
-import org.obiba.mica.file.AttachmentState;
-import org.obiba.mica.file.FileStoreService;
-import org.obiba.mica.file.FileUtils;
-import org.obiba.mica.file.InvalidFileNameException;
+import org.obiba.mica.file.*;
 import org.obiba.mica.file.event.FileDeletedEvent;
 import org.obiba.mica.file.event.FilePublishedEvent;
 import org.obiba.mica.file.event.FileUnPublishedEvent;
@@ -77,11 +56,26 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 
-import com.google.common.base.Strings;
-import com.google.common.base.Throwables;
-import com.google.common.collect.Maps;
-import com.google.common.eventbus.EventBus;
-import com.google.common.eventbus.Subscribe;
+import javax.annotation.Nullable;
+import javax.inject.Inject;
+import javax.validation.constraints.NotNull;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Paths;
+import java.time.LocalDateTime;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.locks.ReentrantLock;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipOutputStream;
+
+import static java.util.stream.Collectors.toList;
 
 @Component
 public class FileSystemService {
@@ -150,10 +144,10 @@ public class FileSystemService {
       saved.setLastModifiedBy(getCurrentUsername());
     }
 
-    if(saved.isJustUploaded()) {
+    if(saved.isJustUploaded() && saved.getId() != null) {
       if(attachmentRepository.existsById(saved.getId())) {
         // replace already existing attachment
-        fileStoreService.delete(saved.getId());
+        fileStoreService.delete(saved.getFileReference());
         attachmentRepository.deleteById(saved.getId());
       }
       fileStoreService.save(saved.getId());
@@ -188,6 +182,11 @@ public class FileSystemService {
    */
   public void delete(AttachmentState state) {
     if(state.isPublished()) publish(state, false);
+    List<Attachment> attachments = attachmentRepository.findByPathAndNameOrderByCreatedDateDesc(state.getPath(), state.getName());
+    attachments.forEach(a -> {
+      attachmentRepository.delete(a);
+      fileStoreService.delete(a.getFileReference());
+    });
     attachmentStateRepository.delete(state);
     eventBus.post(new FileDeletedEvent(state));
   }
